requirements.txt
blinker==1.9.0
chardet==4.0.0
click==8.1.3
cockpit @ file:///builddir/build/BUILD/cockpit-334.2/tmp/wheel/cockpit-334.2-py3-none-any.whl
dasbus==1.4
dbus-python==1.2.18
distro==1.5.0
file-magic==0.4.0
Flask==2.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
gpg==1.15.1
greenlet==3.2.4
idna==2.10
importlib_metadata==8.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
libcomps==0.1.18
MarkupSafe==3.0.3
nftables==0.1
perf==0.1
pexpect==4.8.0
Pillow==10.1.0
psutil==5.8.0
psycopg2-binary==2.9.9
ptyprocess==0.6.0
PyGObject==3.40.1
pyinotify==0.9.6
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==1.0.0
python-linux-procfs==0.7.3
pyudev==0.22.0
PyYAML==5.4.1
requests==2.25.1
rpm==4.16.1.3
selinux==3.6
sepolicy==3.6
setools==4.4.4
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.32/src
six==1.15.0
sos==4.10.0
SQLAlchemy==2.0.44
systemd-python==234
typing_extensions==4.15.0
urllib3==1.26.5
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.23.0
